+++ packages/core/src/components/drawer/test/drawer.spec.tsx
@@ -6,25 +6,8 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-jest.mock('animejs', () => {
-  const mockAnimate = jest.fn(() => ({
-    finished: Promise.resolve(),
-    play: jest.fn(),
-    pause: jest.fn(),
-    restart: jest.fn(),
-  }));
 
-  return {
-    __esModule: true,
-    default: mockAnimate,
-    animate: mockAnimate,
-    timeline: jest.fn(() => ({
-      add: jest.fn().mockReturnThis(),
-      finished: Promise.resolve(),
-    })),
-  };
-});
-
+import './animejs.mock';
 import { newSpecPage } from '@stencil/core/testing';
 import { fireEvent } from '@testing-library/dom';
 import { Drawer } from '../drawer';
