+++ packages/core/src/components/message-bar/test/message-bar.spec.tsx
@@ -7,25 +7,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-jest.mock('animejs', () => {
-  const mockAnimate = jest.fn(() => ({
-    finished: Promise.resolve(),
-    play: jest.fn(),
-    pause: jest.fn(),
-    restart: jest.fn(),
-  }));
-
-  return {
-    __esModule: true,
-    default: mockAnimate,
-    animate: mockAnimate,
-    timeline: jest.fn(() => ({
-      add: jest.fn().mockReturnThis(),
-      finished: Promise.resolve(),
-    })),
-  };
-});
-
+import './../../drawer/test/animejs.mock';
 import { newSpecPage } from '@stencil/core/testing';
 import { fireEvent } from '@testing-library/dom';
 import { MessageBar } from '../message-bar';
